+++ report-viewer/src/views/ComparisonView.vue
@@ -69,8 +69,8 @@ import type { Match } from '@/model/Match'
 
 import { onMounted, ref, watch, type Ref, computed, onErrorCaptured, type PropType } from 'vue'
 import TextInformation from '@/components/TextInformation.vue'
-import MatchList from '@/components/MatchList.vue'
-import FilesContainer from '@/components/FilesContainer.vue'
+import MatchList from '@/components/fileDisplaying/MatchList.vue'
+import FilesContainer from '@/components/fileDisplaying/FilesContainer.vue'
 import { store } from '@/stores/store'
 import Container from '@/components/ContainerComponent.vue'
 import { HighlightLanguage } from '@/model/Language'
